Divided student and teacher hybrid modules into separate files and put common modules in common-hybrid-angular.module file. #2788
