+++ src/main/webapp/site/src/app/student-hybrid-angular.module.ts
@@ -1,0 +1,68 @@
+import { NgModule } from '@angular/core';
+
+import { createStudentAngularJSModule} from '../../../wise5/vle/student-angular-js-module';
+import { UpgradeModule } from '@angular/upgrade/static';
+import { setUpLocationSync } from '@angular/router/upgrade';
+import { ProjectService } from '../../../wise5/services/projectService';
+import { VLEProjectService } from '../../../wise5/vle/vleProjectService';
+import { PossibleScoreComponent } from './possible-score/possible-score.component';
+import { CommonModule } from '@angular/common';
+import { StudentDataService } from '../../../wise5/services/studentDataService';
+import { MatDialogModule } from '@angular/material/dialog';
+import { ChooseBranchPathDialogComponent } from './preview/modules/choose-branch-path-dialog/choose-branch-path-dialog.component';
+import { MatButtonModule } from '@angular/material/button';
+import { MatListModule } from '@angular/material/list';
+import { DataService } from './services/data.service';
+import { AngularJSModule } from './common-hybrid-angular.module';
+
+@NgModule({
+  declarations: [
+    PossibleScoreComponent
+  ],
+  imports: [
+    AngularJSModule
+  ],
+  providers: [
+    { provide: DataService, useExisting: StudentDataService },
+    { provide: ProjectService, useExisting: VLEProjectService },
+    VLEProjectService
+  ],
+  exports: [
+    CommonModule,
+    MatButtonModule,
+    MatDialogModule,
+    MatListModule
+  ]
+})
+export class StudentAngularJSModule {}
+
+@NgModule({
+  declarations: [
+    ChooseBranchPathDialogComponent
+  ],
+  imports: [
+    StudentAngularJSModule
+  ]
+})
+export class StudentVLEAngularJSModule {
+  constructor(upgrade: UpgradeModule) {
+    bootstrapAngularJSModule(upgrade, 'vle');
+  }
+}
+
+@NgModule({
+  imports: [
+    StudentAngularJSModule
+  ]
+})
+export class PreviewAngularJSModule {
+  constructor(upgrade: UpgradeModule) {
+    bootstrapAngularJSModule(upgrade, 'preview');
+  }
+}
+
+function bootstrapAngularJSModule(upgrade: UpgradeModule, moduleType: string) {
+  createStudentAngularJSModule(moduleType);
+  upgrade.bootstrap(document.body, [moduleType]);
+  setUpLocationSync(upgrade);
+}
