+++ src/main/webapp/wise5/common-angular-js-module.ts
@@ -32,6 +32,7 @@ import './components/draw/drawComponentModule';
 import './components/embedded/embeddedComponentModule';
 import * as fabric from 'fabric';
 window['fabric'] = fabric.fabric
+import Filters from './filters/filters';
 import './lib/highcharts/highcharts-ng';
 import * as Highcharts from './lib/highcharts/highcharts.src';
 import './lib/draggable-points/draggable-points';
@@ -115,6 +116,7 @@ import './themes/default/theme';
   .factory('TagService', downgradeInjectable(TagService))
   .factory('StudentDataService', downgradeInjectable(StudentDataService))
   .factory('UtilService', downgradeInjectable(UtilService))
+  .filter('Filters', Filters)
   .config([
       '$httpProvider',
       '$locationProvider',
